initial refactoring to clean up and support multiple outout formats
