+++ Cargo.toml
@@ -9,6 +9,7 @@ edition = "2018"
 [dependencies]
 osmpbfreader = "0.13.4"
 clap = "2.33.3"
+geojson= "*"
 
 [build-dependencies]
 protobuf-codegen-pure = "2"
